Ascii File Node Generator Generalized
The generator will now grab the field names from the first non-header (not following a #) row
and fill the arrays corresponding to those fields. Now the file need not be in any particular order
and as long as it has x,y,z etc in there somewhere, it should just work.