Fiddling a bit with Nick's experimental momentum equation for CSPH. Looking promising so far.  Changes:
1.  There seems to still be a lurking sign error in the force.  I've
arbitrarily flipped it now to check it, but we need to figure out why
this is.

2.  Converted weighting to volume weighting as requested by Nick.  There
are three places this needs to be changed in CSPHHydroBase.cc, all
denoted with the comment "// Change CSPH weights here if need be!".

The acoustic wave solution is definitely in the ball-park with this
form, which is hopeful.