Adding the option of building connectivity for the ghost nodes in ConnectivityMap.  I've also added a new Physics attribute that allows Physics packages to select whether or not ghost connectivity should be computed -- this option defaults to false unless overridden.
The only annoying thing about this change is now we require the ConnectivityMap
be notified as to whether ghost connectivity is required when building or
reconstructing the connectivity.  This interface change touches a fair number of
places in the code (anyone trying to get the ConnectivityMap for instance),
which represents the majority of files touched in this change.