finished icosahedron shape/refinement logic
this should now choose the correct shape and refinement to match a desired shell count automatically.
it is no longer necessary to keep the long resolution table to browse, however, i have left that table in
for the time being since it A) took a long time to create and B) may be necessary in the future.