Switched StrainPorosity velocity gradient
Changed the velocity gradient that StrainPorosity uses to calculate distension time-derivative from the nodelist-local velocity gradient to the nodelist-global version. This should give better consistency w/ the density time-derivative which also uses the global version. The difference seemed to cause some ringing in the distension/density for quasi-steady state interactions of different nodelists on longer run-outs.