improved limiting
limiting based on the GSPH velocity gradient created too much diffusion at the FSI. Changed to using the raw SPH velocity gradient to determine velocity slope limiter and then apply that to the GSPH velocity gradient. Pressure gradient changed to the raw SPH version for now (this will need to be adjusted to deal with flows w/ gravitation).