compatible energy changed to allow opposite sign pairwise DepsDt
For GSPH we need to allow pairwise internal energy DDts to be opposite signs. To do this we use basically the same compatible energy formulation but the difference between the eval-derivs calculated pairwise work (from DepsDti DepsDtj) and the conservative pairwise work (from acceleration) is used instead of the total pairwise work. This difference is subtracted from DepsDti & j to enforce conversation. The quantity subtracted is weighted by magnitude of DepsDti & j