rigorous conservation restored FSI compatibility
compatible formulation requires perfect i-j reversibility? ran into conservation issue when switching i/j would lead to slightly different evaluations of a loop cycle. Manifested as energy error on the first timestep cycle when numeric_limits was applied to the denomenator instead of the individual weights of the compatibility scheme