FSISPH decoupling fix
the decoupling logic in FSISPH was incompatible with the way damage is implemented in the new approach to strength. Once the max eigen of the damage tensor is saturated the particle yields, the deviatoric stress tensor is zeroed, and for many strain models no further damage can accummulate. The decoupling however was based on the directional damage and so large tensile load can develop in fully damaged material.