FSISPH: update deviatoric policy and add interface Pmin
FSI now has the option to run plane strain with a standard update deviatoric stress policy which allows for nonzero trace. Old version of FSI didn't allow tensile loads across interfaces, in this update 'interfacePmin' was added which allows for a non-zero limit to the interface pressure. The thought was interfacePmin was to model adhesion. Finally we have a topical update, the evalDerivs method was shifted over to the .cc file consistent with style changes to the default SPH