DEM conserving w/ redistribution now
I played it a little too loose w/ the mechanics of the pairwise storage in the initial release of DEM. It was not i-j flip safe -- i.e. for a particle pair, if the i-j ordering got flipped the stored values were actually the opposite of what they should have been. With these edits, we transform to a partition / i-j independent prefered orientation'' to store the pairwise values. In this preferred orientation, i is the node with the min unique index. In the evaluateDerivs this leads to some -1 : 1 complications which we should buff out at some point. Rolling friction wasnt affected sinces its i-j orientation independent. Redistribution is working but kullFreq>1 still gives a conservation issue.