DEM fix for KullFrequency consevation bug
Okay, so before we would break conservation if our periodic cleaning of old contacts happened at a Frequency > 1. Weirdly you'd be fine if the kullFrequency was effectly infinite meaning if you never cleaned. This turned out to be an issue with how the integrator interacts with the physics object. Because we had this cleaning in the initialize method it could potentially happen mid cycle within the integrator. This meant the state and derivs could get out of sync with predictor-corrector algos.  In this edit we move the cleaning to the preStepInitialize method to make sure the contactMap doesn't change mid-cycle anymore. NOTE: I don't think DEM is compatible with the RigorousBoundaries flag in the integrator.