DEM: particle-solid bc interactions added to eval derivs and moment of inertia reworked
this commit added the logic for keeping track of the particle - solid bc interaction in the contact map as well as pair loop over the particle-solid bc interactions to calc the derivs. The moment of Inertia was give its own fieldList in the LinearSpring class. I avoided putting it in the base class b/c i think for non spherical particles we might want the MOI to behave differently.