Make the Spheral license detectable
GitHub uses [Licensee](https://github.com/licensee/licensee) to
detect licenses and Spheral doesn't show up as BSD 3-clause
(even though it is) because it is adding a bunch of text to the
license file and appending the lab's additional notice, which
can be separate.

See https://software.llnl.gov/about/licenses/ for more.

- [x] Use the standard BSD 3-clause license format
- [x] Use the LICENSE
- [x] Separate the NOTICE file
- [x] Spruce up the README and add a SPDX identifier